refactor: Move more general ApiItem utilities
