+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiClass.ts
@@ -14,10 +14,9 @@ import {
 } from "@microsoft/api-extractor-model";
 
 import { SectionNode } from "../../documentation-domain";
-import { filterByKind, isStatic } from "../ApiItemTransformUtilities";
+import { ApiModifier, filterByKind, isStatic } from "../../utilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
 import { createChildDetailsSection, createMemberTables } from "../helpers";
-import { ApiModifier } from "../../utilities";
 
 /**
  * Default documentation transform for `Class` items.
