+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiEnum.ts
@@ -5,7 +5,7 @@
 import { ApiEnum, ApiEnumMember, ApiItem, ApiItemKind } from "@microsoft/api-extractor-model";
 
 import { DocumentationNode, SectionNode } from "../../documentation-domain";
-import { filterByKind } from "../ApiItemTransformUtilities";
+import { filterByKind } from "../../utilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
 import { createMemberTables, wrapInSection } from "../helpers";
