+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiModuleLike.ts
@@ -15,10 +15,9 @@ import {
 } from "@microsoft/api-extractor-model";
 
 import { SectionNode } from "../../documentation-domain";
-import { filterByKind } from "../ApiItemTransformUtilities";
+import { ApiModuleLike, filterByKind } from "../../utilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
 import { createChildDetailsSection, createMemberTables } from "../helpers";
-import { ApiModuleLike } from "../../utilities";
 
 /**
  * Default documentation transform for module-like API items (packages, namespaces).
