+++ tools/api-markdown-documenter/src/index.ts
@@ -25,20 +25,9 @@ export {
 	doesItemRequireOwnDocument,
 	// TODO: remove this once utility APIs can be called with partial configs.
 	getApiItemTransformationConfigurationWithDefaults,
-	getDefaultValueBlock,
-	getDeprecatedBlock,
-	getExampleBlocks,
 	getHeadingForApiItem,
 	getLinkForApiItem,
-	getModifiers,
-	getReleaseTag,
-	getReturnsBlock,
-	getThrowsBlocks,
 	type HierarchyBoundaries,
-	isDeprecated,
-	isOptional,
-	isReadonly,
-	isStatic,
 	type TransformApiItemWithChildren,
 	type TransformApiItemWithoutChildren,
 	transformApiModel,
@@ -84,9 +73,20 @@ export {
 	type ApiModifier,
 	type ApiModuleLike,
 	type ApiSignatureLike,
+	getDefaultValueBlock,
+	getDeprecatedBlock,
+	getExampleBlocks,
+	getModifiers,
 	getQualifiedApiItemName,
+	getReleaseTag,
+	getReturnsBlock,
 	getSeeBlocks,
+	getThrowsBlocks,
 	getUnscopedPackageName,
+	isDeprecated,
+	isOptional,
+	isReadonly,
+	isStatic,
 } from "./utilities";
 
 // #region Scoped exports
